Dont't dispatch onPageScrolled without pages
When the ViewPager2 does not have an adapter set, you can still perform
drag gestures. RecyclerView will change the scroll state to dragging and
back to idle after the drag is complete, but won't issue scroll events
because there is nothing to scroll in the first place.

Our ScrollEventAdapter now thinks it is at the start or end of the list
because there were no scroll events during the drag, and issues an
onPageScrolled event. Prevent this from happening by checking if the
position for the onPageScrolled event is valid before dispatching the
event.

Bug: None
Test: ./gradlew viewpager2:connectedCheck
Change-Id: Ibafa7fcd532292e90a48db46698cb7094bc38973