Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The problem is that there's no action for the developer here, there's nothing they do. We want to reduce logcat spam as much as possible.
> 
> It's only useful for us, and we can just debug in AS

It's useful for an app developer too - since this is hitting non-public API, it's theoretically possible that there's some device that modified these APIs. If a developer is seeing weird behavior on that device, their first guess probably wouldn't be that recreate() is broken somehow - on the other hand, if they see this message, they will have an idea of what is going wrong. If we change this to only fail if neither method is found, then it will not print anything to logcat on any device we know of, and thus not be logcat spam.

Patch-set: 1