Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thinking about it more, I'm not sure why we (WorkManager) need to do anything to facilitate custom threading testing.  This is something that the developer should be expected to do.

I was in 2 minds about this as well. The developer can technically use LiveData<> and block until a Worker has reached a terminal state, but these helpers make it a lot easier because they don't have to write boilerplate code for it.

Also, for PeriodicWork - we can mark it complete on every run of the Worker - so they can drain for every run of the PeriodicWorkRequest.

Patch-set: 1
Reviewer: Gerrit User 1344433 <1344433@85c56323-6fa9-3386-8a01-6480fb634889>