Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Thinking about it more, I'm not sure why we (WorkManager) need to do anything to facilitate custom threading testing.  This is something that the developer should be expected to do.
> 
> I was in 2 minds about this as well. The developer can technically use LiveData<> and block until a Worker has reached a terminal state, but these helpers make it a lot easier because they don't have to write boilerplate code for it.
> 
> Also, for PeriodicWork - we can mark it complete on every run of the Worker - so they can drain for every run of the PeriodicWorkRequest.

I don't think that's a good idea.  If we're changing the behavior of WM just to accommodate a test they can write themselves, I don't think we should do it.

Moreover, something like drainAll() strongly implies there is no previously-enqueued work, and that WM is in a hermetic state, which it may not be.

Patch-set: 1
Reviewer: Gerrit User 1060206 <1060206@85c56323-6fa9-3386-8a01-6480fb634889>