DO NOT MERGE - Fix the preivew of CoreTestApp will become black screen when press home key and launch again
Root cause:
If activity sets OnSurfaceTextureListener to TextureView and return "false"
via onSurfaceTextureDestroyed() callback, TextureView won't automatically
release the SurfaceTexture when it leaves, but when returning to activity
, SurfaceTexture is no longer attached to to TextureView, hence causing
black screen.

Solution:
So far, there is no suitable signal or event to drop the SurfaceTexture
when pausing activity. Use a workaround to create a new SurfaceTexture
when creating a capture session everytime. By the existing flow, it will
callback SurfaceTexture and attach it to TextureView.

Bug: 128487010
Test: Manually test.
Change-Id: Ic3a2a22944213dbd69ce7238424934573665c9d8