Support Kotlin Coroutines Channel & Flow as query return types.
This CL adds QueryResultBinders and Providers for both Channel and
Flow. Similar to Rx the strategy is to generate code that calls
into a runtime helper methods to setup the Channel or Flow given
a Callable that actually executes the query.

DAO functions that return SendChannel or Channel are an error.
Suspend DAO functions that returns these two types are also
an error.

Bug: 127328278
Test: CoroutinesRoomTest, ChannelQueryTest, FlowQueryTest
Change-Id: If0fc6fdb741735f2e4b6b42015730bbf76e8dc18