Update patch set 10
Patch Set 10:

This presubmit failure looks to me like the same one we've been seeing recently.

This change doesn't affect the contents of the libraries so I'll submit it anyway.

I also opened b/130562688 for making it easier to determine when to open a new bug for a presubmit failure.

Patch-set: 10
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>