Restructured extensions in order to fix vendor lib violation
Separated out dependencies that the OEM library needed for creating the
extensions library into extensions-stub. This includes the interfaces
and some of the convenience classes.
This solves the circular dependency issue where the OEM library was
loading classes from the application. This was a violation of Q policy
which manifested as a NoClassFoundError.

Bug: b/129418078
Test: ./gradlew camera:camera-extensions*:build, manual test on device
Change-Id: I806b924f65021d39b0cbb4737ed20b4d63865d0f