Activity launching benchmark runner
Bug:111923337
Bug:129282573

Adds a (currently optional) benchmark runner to the benchmark library, which when used launches an
'isolation activity' to ensure our benchmarks are running on top, without the launcher or sysui
showing.

Currently, there's no label on the benchmark activity, so device inexplicably blinks black.

This work is a prerequisite to using Window.setSustainedPerformanceMode on devices that support it.

Test: AndroidBenchmarkRunnerTest

Change-Id: Id9dd5471ebd02b0639e46ee450568661b8da5f2e