Save fragment mRemoving on save and restore
The mRemoving flag is not saved when a fragment goes through save
state. This means when a configuration change (or any other save and
restore cycle), the flag is lost. If a fragment is on the backstack,
this can cause that fragment to never reach the CREATED state again.

Adding the mRemoving flag to FragmentState ensures that fragments on
the backstack get to the CREATED state. A fragment can now also get to the
DESTROYED state if it's FragmentManager is destroyed, meaning the
non-config state (viewModels) is properly cleared.

Test: new ViewModelTest, new SaveStateFragmentTest, rerun old test
Fixes: 129593351
Change-Id: Ia1d10e96f569dee606de2ac19c586901b240b3cd