Update patch set 1
Patch Set 1:

The confusing part of this change is that if you bind a use case no only could you possibly stop use cases that have been bound to another lifecycle, but if there exists use cases that are bound to the same lifecycle to start up.

And as mentioned in my comment, this policy decision should not be built into the UseCaseGroupLifecycleController. The single job of the UseCaseGroupLifecycleController should be controlling that single UseCaseGroup.

Patch-set: 1
Reviewer: Gerrit User 1482448 <1482448@85c56323-6fa9-3386-8a01-6480fb634889>