Update patch set 2
Patch Set 2:

After consult with Wenhung@, it doesn't have to release the surface in StateCallback.onConfigured() because in normal flow it will call session.close() in StateCallback.onConfigured() and lead to surface be released in StateCallback.onClosed().
Test this change (patch version 1->2) on legacy devices and did not see the StrictMode violation.

Patch-set: 2
CC: Gerrit User 1475857 <1475857@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1481635 <1481635@85c56323-6fa9-3386-8a01-6480fb634889>