Not to remove the same animator twice
Transition.forceToEnd may be called during the execution of method
itself. In that case, an item of the map was removed during the
previous iteration and ArrayIndexOutOfBoundsException was raised.

This CL updates the loop to iterate the copy of map instead of the
original map instance, so we don't break the iteration and see
exception anymore.

Bug: 120675604
Test: Ran 'cts-tradefed run commandAndExit cts --module
    CtsWidgetTestCases --test android.widget.cts.ToolbarTest
    #testMenuOverflowSubmenu' 20 times and no failure was observed.

Change-Id: I95b6c6c3b60d4aeaf2a07c2cb2de8f50797a3416
(cherry picked from commit 8b1665dd70b2c3bda327d44d8228b60f0dd3c399)