Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 2:
> 
> (1 comment)
> 
> +etalvala who's had a fair bit of experience with camera sidecar libraries.
> 
> Main concerns:
> - How will be handle versioning?
> - How will be handle missing or only partial implementation of the api surface?
> - How will be test the there API implementation in OEM devices?

Those issues are not handled in the current change, but to answer:
Versioning - The plan is using rules of semantic versioning for checking version compatibility. If needed I can add a List<int> getVersion(), where [0] is major, [1] is minor, [2] is patch to this change.

Partial/missing implementation - Extensions allows for individual Impl types to be unimplemented. For example, if BokehImageCaptureExtenderImpl is partially implemented then cameraX will replace the implementation with a no-op implementation.

Test for API implementation - Not sure if this is the answer to your question, but we utilize <uses-library> to load the implementation.

see go/camerax-extensions-versioning for more comprehensive details.

Patch-set: 3
Reviewer: Gerrit User 1482448 <1482448@85c56323-6fa9-3386-8a01-6480fb634889>