Update patch set 7
Patch Set 7:

(1 comment)

As this is a testing dependency, I'd suggest writing this new API in Kotlin and pull in the Kotlin standard library for the testing artifact. This is the case for fragment-testing and is approved across AndroidX for testing libraries.

Patch-set: 7
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>