Update patch set 1
Patch Set 1:

I think printing ns is already info overload. It's already really hard to read milliseconds. Having two numbers in ns is too much, and probably makes it hard to line the times up with names. In future patches, could you paste sample output into the commit message, so we can verify?

I don't think it's disingenuous - we report the min observed, but if we observe a certain amount of variation, it's reasonable to expect the minimum possible could be lower than our minimum observed.

Also, we can consider rejecting some of the data when we're making our calculation - remember part of the reason of reporting the min is assuming the higher numbers are corrupted by interference, and invalid. Maybe say +- with a percentage that's = `(2nd highest) / minimum`

I think it's not disingenuous, since we're not quoting some statistical term - we're just trying to show vaguely how confident we are in as few characters as possible.

Patch-set: 1
Reviewer: Gerrit User 1010334 <1010334@85c56323-6fa9-3386-8a01-6480fb634889>