Update patch set 1
Patch Set 1:

> But the variance we observe is calculated from the effects of interference on the mean, that number doesn't really say anything about the min, except how unlikely the min is to occur.
> 
> One thing we could do is report how many standard deviations the min is from the mean?

I like the idea of reporting how far apart the mean and min are, though maybe median is better to avoid including outliers at the high end.

If we avoid standard statistical naming, so that we're not implying this is stddev or variance or whatnot, then we could represent that value as a percent (just to make it fewer characters). We just need to think of a name.

Patch-set: 1
Reviewer: Gerrit User 1010334 <1010334@85c56323-6fa9-3386-8a01-6480fb634889>