Add isValidDeepLink Api to NavController
When users attempt to navigate to a deep link destination, if the deep
link is invalid an IllegalArgumentException is thrown. This Api offers
a way to verify the deeplink is valid before attempting to navigate, and
gives developers more flexibility to handle the case where the deep link
may be invalid.

It uses NavGraph's matchDeepLink() and returns whether a match was
found.

Test: New tests and ./gradlew checkApi
BUG: 117437718
Change-Id: Iabbca9cfbb864ce305d84d9e4f616d2b43dc7ac6