Update patch set 4
Patch Set 4: -Code-Review

I think we're trying to make too many consolations to accommodate the edge cases of TextureView. If the user decides to return 'false' from OnSurfaceTextureListener, then it should be the user's responsibility to re-attach the SurfaceTexture once returning to the activity. By returning 'false', the user is signalling that they don't want to release those resources, so we shouldn't be recreating the resources for them.

Patch-set: 4
Reviewer: Gerrit User 1377888 <1377888@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0