Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> We had talked about having this as NavDestination#hasDeepLink(Uri) instead of something on NavController and encouraging devs to do navController.graph.hasDeepLink(uri) since that would nicely mirror the addDeepLink API already on NavDestination. Any thoughts on that approach?

That part slipped my mind while writing it. Sorry about that!

I agree that having it in NavDestination would probably be the best place considering that addDeepLInk is already there and these two go hand and hand. With navigate(Uri) and handleDeepLink(Intent) being on NavController, needing to access the graph for a check does seem a little unnatural, but with the right communication I believe it will be okay.

Thoughts on foundDeepLink(Uri) or isDeepLinkAdded(Uri) for the Api name?
hasDeepLink(Uri) makes me think "has a deep link" instead of "has this deep link". The parameter is a big hint, but that is where my mind goes.

Patch-set: 2
Reviewer: Gerrit User 1442279 <1442279@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0