Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> Looks great, but you need a test to validate that this works and we don't break it in a future change.

I've just found that this is hard to write a test to reproduce this. This crash is depending on the hash values of the animation object. This happens only when the animation object of the parent menu is larger (in their hash values) than the one of child, and the parent object gets ended earlier. We can't determine object's hash values in the test:

And I wrote the summary of this crash here:
https://issuetracker.google.com/issues/120675604#comment8

Patch-set: 2
Reviewer: Gerrit User 1087058 <1087058@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0