Make a node a heading with setHeading API
Use the newer API since the heading parameter of
CollectionItemInfoCompat.obtain is deprecated. Developers will be able
to change this property on an item view themselves now if they don't
like the layout manager's label.

The Movies app is facing an issue where one of their items that spans
the screen is not a heading but is labeled as such by GridLayoutManager, so this
will give it  a way to relabel

Bug: b/117202844
Test: Passed RecyclerViewAccessibilityTest and
RecyclerViewAccessibilityLifecycleTest
Change-Id: Ibc6880ea57acf5a3d523137dffb02dc8c564d38c