Use IntDef instead of boolean in FragmentPagerAdapter
Deprecated the orignal constructors for FragmentPagerAdapter and
FragmentStatePagerAdapter to encourage developers to use the new
constructor that allows the behavior of only allowing the current
fragment to be in a RESUMED state.

Changed the new constructor to take an IntDef instead of a boolean so
that the intended behavior is clear. Passing in
USE_SET_USER_VISIBLE_HINT uses the old behavior and passing
RESUME_ONLY_CURRENT_FRAGMENT uses the new behavior.

Test: support v4 demo, support v7 demo, ./gradlew checkApi
BUG: 129780800
Change-Id: Ie2fcf318424fd4037ff6c541ba5a6e901c97fc5b