Account for delays when sending drags in FakeDragTest
In FakeDragTest, calls to ViewPager2's fake drag methods are scheduled
such that it results in a smooth drag. But the actual time at which
those callbacks are executed is unreliable (too late), which influences
the velocity calculated by Viewpager2. Sometimes this results in a
velocity above the fling threshold when a velocity below the threshold
was expected, which causes a fling to happen when it shouldn't.

Fix this by calculating the offset passed to fakeDragBy dynamically,
when the callback is executed. That makes the graph described by
offset = f(time) is always smooth, even when the time between points is
irregular.

Bug: 129296837
Test: ./gradlew viewpager2:connectedCheck
Change-Id: Ib0a8f057a80ea20280029899545691e4c78e2669