Fix child fragment animation issues
When a child fragment is on top of the parent and that parent is replaced,
the child fragment's animation from the most previous operation is
executed. This can cause unexpected behavior in developer animations.

We should clear animations after they have been executed and child
fragments with no animation of their own, should inherit the animation
of their parent.

Test: Tested visually in app
BUG: 121017790
Change-Id: I4f008316ea00258f9bec445216f1b612330d509a