Update patch set 1
Patch Set 1: -Autosubmit Presubmit-Ready+1

Crud, passing a null in the parameterized test didn't work.. on API 17 for some reason.  Removed the null assuming that "doing_nothing" PostLayoutRunnable suffices to test the intended path.

Is everything that was originally intended to be tested still adequately tested with this change?

Patch-set: 1
Label: Autosubmit=0
Label: Presubmit-Ready=+1