Update patch set 1
Patch Set 1:

(1 comment)

Why is it the case that we need to change the interface of the builder to add base parameters?

It seems it would be cleaner to simply apply the implementation options for both the repeating parameters and the single request parameters.

Also what happens if there are parameters in the repeating request that want to override? Is there a way to do so?

Patch-set: 1
Reviewer: Gerrit User 1482448 <1482448@85c56323-6fa9-3386-8a01-6480fb634889>