Add instrumented tests to Antelope
Instrumented tests:
 - basic context sanity check
 - single capture
 - multiple capture w/ abort
 - single capture saving and deleting photo
 - switch camera back->front->back test
 - full testing suite for all apis/cameras/image sizes/tests

Currently tests are marked @FlakyTest so as not to block
all of androidx.

To test, use an emulator (from a non-canary Android Studio install - not
from the androidx directory) or a connected device. For best results,
disable animations in developer options on the device. If a test fails
due to not being able to find a button, this is probably due to flakey
back-button processing to close the settings dialog, just re-run the
test.

Pixel2 XL, API 28 used as a reference.

Test: gradlew camera:integration-tests:camera-testapp-timing:connectedCheck --info --daemon
Change-Id: I3dca5ff789bbfb9bf5d6cafb07df2b084dcc0591