Add instrumented tests to Antelope
Instrumented tests:
 - basic context sanity check
 - creating and deleting log files
 - single capture
 - multiple capture with front camera
 - multiple capture chained
 - multiple capture w/ abort
 - switch camera back->front->back
 - single capture saving and deleting photo
 - multiple capture with Camera1
 - multiple capture with Camera2
 - full testing run for all cameras/image sizes/tests with Camera2
 - full testing run for all cameras/image sizes/tests with CameraX
 - full testing run for all cameras/image sizes/tests with Camera1

To test, use an emulator or a connected device. Note: switch camera
tests will not run if device does not have both a back and a front
camera. Most tests will be skipped if device does not have any cameras.

Pixel 2 XL and Pixel 3 on API 28 were used as reference devices.

Test: gradlew camera:integration-tests:camera-testapp-timing:connectedCheck --info --daemon

Change-Id: I3dca5ff789bbfb9bf5d6cafb07df2b084dcc0591