Address flakiness in FakeDragTest
Addresses several AssertionErrors that arise every now and then in
FakeDragTest, mostly on a Nexus 4 running API 17. The common issue was
that the test process stuttered (lag of ~40 till 300ms), due to which
test actions were timed incorrectly.

Fixed by:
1) starting test actions on the UI thread
2) checking some preconditions before starting those actions
3) reset viewPager and try again if preconditions not met
4) adjusted some parameters to eliminate the role of VelocityTracker

A list of AssertionError messages that are fixed by these changes:
- "Last event should be state change to IDLE"
- "viewPager.getCurrentItem() should return x"
- "When a fake drag should end in a snapped position, we expect the last
  scroll event after the FAKE_DRAG event to be snapped."

Bug: 131579825
Bug: 131581454
Bug: 131583038
Test: ./gradlew viewpager2:cC

Change-Id: I62e545da4c891888305d137fcaa54cf6974533a1