Ensure parent fragments consider children when handling menus
There is currently no way for a parent fragment to know about menus in a
child fragment. This can create situations where menus are not handled
correctly (i.e. a parent Fragment is removed, but the child's menu
remains).

This change allows for parent fragments to check its children for menus
and invalidate the menu properly. This ensures actions performed on the
parent Fragment can reflect appropriately in the menus associated with
that parent, as well as its children.

Test: Added OPtionMenuFragmentTest, reran Fragment Test
BUG: 131581013
Change-Id: I826cc11d989ccaa7dbf55561c95dbdb63dcbf86c