Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (2 comments)
> 
> > Patch Set 2:
> > 
> > (2 comments)
> > 
> > What happens if there are 2 LifecycleOwners say A & B
> > 
> > 1) A goes to a STARTED state (A is active)
> > 2) B goes to a STARTED state (B is active, A is inactive)
> > 3) B goes to a CREATED state (A is active)
> > 
> > I don't think the controller logic does this. Instead in step 3 A would remain inactive. Or is that the intended behavior?
> 
> I did not change any lifecycle behavior in this CL. My changes still depend on originally lifecycle behavior. If getting one lifecycle ON_START event, stopping use cases belong to other lifecycles.
> 
> In step 3 of your example, if A remain inactive, the UseCaseGroup belong to it won't be start. If it becomes active, ON_START event should trigger the corresponding UseCaseGroup to start. There is no specific intended behavior. The behavior depends on what the lifecycle is. Activity, fragment or customized lifecycle may have different results.

It actually does change behavior. Because with this change if LifecycleOwner B goes to ON_START it will cause A to become inactive. However previously it stays active.
The problem is that the only way for A to become active again it needs to go to ON_CREATE then go back to ON_START.

The issue we are trying to solve is if an Activity goes into the background it doesn't actually go into the ON_DESTROY state so the use case is still bound. So does it make it simpler if ON_CREATE will make the use case go into an inactive state? Instead of only the latest ON_START use case will be active.

Patch-set: 4
Reviewer: Gerrit User 1482448 <1482448@85c56323-6fa9-3386-8a01-6480fb634889>