Update patch set 1
Patch Set 1:

I assume this is to unblock submissions? It would be much better to skip the test instead of rewriting the test to expect the wrong values. Otherwise it would make it more confusing later on what this test was supposed to be about.

But as Sam suggested we should create a bug to track this so we can figure out what is going on, whether the original test is had some logic error, or we made some mistake in our pipeline.

Patch-set: 1
Reviewer: Gerrit User 1482448 <1482448@85c56323-6fa9-3386-8a01-6480fb634889>