Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> FTR, we had this discussion over chat:
> 
> 1.)

As it stands the most common usage would be:
KeyedAppStatesReporter reporter = SingletonKeyedAppStatesReporter.getInstance();
// pass around and use reporter as a KeyedAppStatesReporter

a rarer usage would be:
SingletonKeyedAppStatesReporter.initialize(...);
KeyedAppStatesReporter reporter = SingletonKeyedAppStatesReporter.getInstance();
// pass around and use reporter as a KeyedAppStatesReporter

so if we moved the getInstance to the base class the first would be simpler, and the second would be awkward unless we also had the getInstance on SingletonKeyedAppStatesReporter too.

I guess I'll add it to the base class without removing it from Singleton* as that doesn't seem to have many downsides.


> 2.) 

I don't think anything outside of the package should have to subclass it. The fake will be in the main src directory with its own tests in the test directory, and all tests should use that fake so there should be no reason for anyone to extend the base class themselves.

I don't suppose there are any big problems with `/** @hide */ @TestApi protected` instead - though it does mean there's nothing to stop clients from extending the base class rather than using the provided Fake which means there could be more dependencies on the base class than would be ideal.

If it was package-private, would that prevent it from being changed to `/** @hide */ @TestApi protected` in future if the disadvantages you listed became important?

Patch-set: 8
Reviewer: Gerrit User 1411450 <1411450@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0