Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > FTR, we had this discussion over chat:
> > 
> > 1.)
> 
> As it stands the most common usage would be:
> KeyedAppStatesReporter reporter = SingletonKeyedAppStatesReporter.getInstance();
> // pass around and use reporter as a KeyedAppStatesReporter
> 
> a rarer usage would be:
> SingletonKeyedAppStatesReporter.initialize(...);
> KeyedAppStatesReporter reporter = SingletonKeyedAppStatesReporter.getInstance();
> // pass around and use reporter as a KeyedAppStatesReporter
> 
> so if we moved the getInstance to the base class the first would be simpler, and the second would be awkward unless we also had the getInstance on SingletonKeyedAppStatesReporter too.

initialize() should only be called by tests, never be called by apps, right? I care more about the cognitive load that we're putting on apps than the cognitive load that we put on tests. (Also, I don't follow how it'd be more awkward, but probably doesn't matter).

> I guess I'll add it to the base class without removing it from Singleton* as that doesn't seem to have many downsides.

Sounds good. The main downside of having equally-named static methods on different classes is that which method gets called is resolved *at runtime*. But as long as they do the same thing, it's fine. You could e.g. call the SingletonKeyedAppStatesReporter.getInstance() method from KeyedAppStatesReporter.getInstance() method to make sure they do the same thing (the subclass method's compile-time return type can be the subclass.

> > 2.) 
> 
> I don't think anything outside of the package should have to subclass it. The fake will be in the main src directory with its own tests in the test directory, 

Why would the fake be in the main src directory? It's only useful for tests, right? We don't ship test utility code (that is only useful for tests) in the Android platform. 

> If it was package-private, would that prevent it from being changed to `/** @hide */ @TestApi protected` in future if the disadvantages you listed became important?

I *think* not, except that it provides the power of setting a (potentially inappropriate) precedent. I'll leave this bit to your judgement, as long as you keep the option open of changing it to protected in future.

Patch-set: 8
Reviewer: Gerrit User 1103203 <1103203@85c56323-6fa9-3386-8a01-6480fb634889>