Update patch set 3
Patch Set 3: Code-Review+1

LGTM, but Room 2.1.0 is in beta and a change in retention is an API change. Once Room 2.1.0 moves to RC (about a week from now or so) we'll have a release branch and head will become 2.2.0-alpha, by then we can merge this.

Patch-set: 3
Reviewer: Gerrit User 1258207 <1258207@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1