Update patch set 8
Patch Set 8:

> initialize() should only be called by tests, never be called by apps, right? I care more about the cognitive load that we're putting on apps than the cognitive load that we put on tests. (Also, I don't follow how it'd be more awkward, but probably doesn't matter).

initialize() would be called by apps which have specific requirements about executor lifecycles. It wouldn't be used by tests. It's similar to the initialize method in WorkManager (which also allows changing the executor).

> > I don't think anything outside of the package should have to subclass it. The fake will be in the main src directory with its own tests in the test directory, 
> 
> Why would the fake be in the main src directory? It's only useful for tests, right? We don't ship test utility code (that is only useful for tests) in the Android platform. 

FTR as discussed over chat: It is only useful for tests, but not our own tests - it's for third party apps who wish to report states and need to test that they are reported in the correct place. The alternative is for every third party app to create their own fakes or mocks - which will be less well tested then ours.

Patch-set: 8
Reviewer: Gerrit User 1411450 <1411450@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0