Update patch set 12
Patch Set 12:

> 2.) KeyedAppStatesReporter currently has a default constructor (public, no-args). You probably want to make it protected. Depending on whether you think that (a) only platform tests, or (b) also apps, should be able to subclass it, you can make it either:
> (a) /** @hide */ @TestApi protected
> (b) protected
> Your choice.

Sorry, when I wrote this I was confused (had forgotten that this here is androidx so one doesn't use @hide but at most @RestrictTo or similar. Since this is androidx, and given that Singleton* is for use by apps (so probably not appropriate to @RestrictTo), perhaps we should have the getInstance() only on the Singleton* after all (drop the one on the base class)?

Does this make sense? (Sorry again, I'm not actually very experienced with androidx, so I think I gave you the wrong advice here)

Patch-set: 12
Reviewer: Gerrit User 1103203 <1103203@85c56323-6fa9-3386-8a01-6480fb634889>