Make GestureDetector listen directly to RecyclerView.
This detangles the event handling a bit. Was using convoluted logic to determine when to delegate events to GestureDetector. After experimentation it didn't appear that convoluted arrangement was necessary.
Added new test for ToolHandlerRegistry.
Rename TouchEventHandler > EventHandler to avoid confusion. Can't convince git this is a rename.  Lower similarity threshold works `git log -M40 -p --stat`, but I'm not sure how to convince gerrit to use the same threshold.

Test: Added new test for ToolHandlerRegistry.

Change-Id: I1d8cc5beb1d3e0eaeec7ab6dfd74809903af5411