Update patch set 1
Patch Set 1: Presubmit-Ready+1

IIRC, the reason we did not flip the default is that it broke a bunch of IDE unit tests in our internal repository, and it wasn't clear to me how to run the IDE tests with the KTX configuration flag on.  Since we're migrating away from KTX, and I think we didn't end up copying those tests over when we migrated into androidx, probably this change is fine.

Patch-set: 1
Reviewer: Gerrit User 1383730 <1383730@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1