Allow NavHosts to enable/disable OnBackPressedCallback
Ensure that a NavHostFragment can correctly control
the enabled/disabled state of the NavController's
OnBackPressedCallback so that it remains in sync
with whether the the NavHostFragment is the primary
navigation fragment.

Importantly, enabling/disabling the callback is
strongly preferred over removing/re-adding the callback
as it ensures that the lifecycle ordering remains
consistent between the NavController and child
FragmentManagers.

By making it a return object from the
setHostOnBackPressedDispatcherOwner(), the control
is restricted to the NavHost only (and those the
NavHost wants to make it available to) rather than
to all accessors of the NavController.

Test: new OnBackPressedTest
Fixes: 132077777
Change-Id: I2bcc33bbf610e155e6bc4c14ba3106b57d0d6418