Update patch set 9
Patch Set 9:

The DeferrableSurface class is becoming too complicated to use. It isn't actually clear anymore what it means to be a DeferrableSurface. In addition, CaptureSession shouldn't need to know that the DeferrableSurface has surfaces it needs to free up. By giving CaptureSession this responsibility it is leaking the details of how what DeferrableSurface is doing. DeferrableSurface really should just be a way for the Surface to be set after the attachToCamera has been called.

I suggest that DeferrableSurface gets simplified so it only has a getSurface() and then some ref counting for releasing. In fact maybe getSurface() should be callable only once (since we are adding the DeferableSurface to Surface map).

Then the question is how to refresh or recreate the Surfaces?
There has already been some offline discussions on how we need the camera2 stack to report errors to the use cases. So in the LEGACY devices we can report an error that the Surfaces need reconfiguration (or something to that effect). Then the UseCase can create new DeferrableSurfaces and send it down.

This will serve the purpose of cleaning up the DeferrableSurface and it will report errors (which really are workarounds) through a consistent mechanism.

Patch-set: 9
Reviewer: Gerrit User 1482448 <1482448@85c56323-6fa9-3386-8a01-6480fb634889>