Implemented saving and restoring of the refreshing state of SwipeRefreshLayout.
In the BaseSwipeRefreshLayoutActivity the Runnable is placed inside
the ViewModel for proper operation after rotating the screen.
build.gradle updated to allow use of the ViewModel in the BaseSwipeRefreshLayoutActivity class.

Test: Open the application support-v4-demos.
Click on the list item "SwipeRefreshLayout ListView".
Perform "swipe-to-refresh" gesture and immediately turn the screen.
After turning the screen, the "Refreshing" indicator should reappear,
and then, after a short time, the "Refreshing" indicator should disappear.
Change-Id: I7229feece86f8776ae8fe7ff368b8b8672bc4b44