Update patch set 4
Patch Set 4: Code-Review+2

> Patch Set 4: Presubmit-Ready+1
> 
> Take a look at patch set #4 and see if that logic makes sense to you

It does. It avoids adding the grandparent to the back stack in the case I mentioned above. Looking at it closer, even if a user explicitly uses the grandparent to navigate, I do not think we need to keep it on the back stack. Looks good to me.

Patch-set: 4
Reviewer: Gerrit User 1442279 <1442279@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2