Introduce Recomposer in favor of CompositionContext
The only purpose of CompositionContext was to manage recomposition of
one composer. In the world, where we can create multiply composition
threes and subcomposition trees, we need the way to manage this is one
place. By doing this, we will centralize the changes dispatching
timings, e.g. control how many composers to recompose in each frame and
what's the order will be.

This CL also adds `hasPendingChanges(): Boolean`  API, which makes sure
there's no invalidations scheduled for composers in the thread. This allows us to wait for idle state
in test infrastructure, which makes much more sense, then waiting for
recompose (because there might be no recompose expected) or waiting fre
frames. This CL also adds related test infra changes.

Bug: b/131238759
Bug: b/132701222
Test: ./gradlew ui-material:connectedCheck, also tests for Recomposer will be soon

Change-Id: I2523a8ed377a426aab670ebc212abc2d21f37271