Update patch set 5
Patch Set 5: Code-Review+1

(1 comment)

Not necessarily part of this change, but it'd be neat if there were some way to make it clearer how many tasks were created during the build so that we accidentally make a change in the future that makes these tasks be created eagerly again then it'd be easier to notice. For example, maybe at the end of the build logging the number of tasks that were created but that weren't a part of the task graph.

Patch-set: 5
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1