Update patch set 18
Patch Set 18:

According to the logs, the tests are completing successfully, but the idling resource does not seem to be detecting that the "single test" button is re-enabled after the test on API 26, and therefore is timing out. This is the signal I am using to determine when a test run is complete.

I will investigate further.

Patch-set: 18
Reviewer: Gerrit User 1269804 <1269804@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0