Change vectordrawable-animated package name.
Package name for a library only affects the package of R.java class.
In the namespaced resource world each R.java file has to be in a unique
package. We were violating that in vectordrawable-animated library.
We got lucky that this library has no resources, so we are not breaking
anyone.

This change also enables the check to make sure we don't make this
mistake again.

Bug: 132781141
Test: ./gradlew assembleDebug passes w/ this patch and check enabled
Change-Id: Id97a6b6bdfe4627826aec65716dd8c8bacf0fe8f