Update patch set 1
Patch Set 1: Code-Review-1

I'm okay with this, but should it go into a release branch? I'm okay with leaving it in master but if you want to add any features it'll have to wait a month for stable to launch.

Patch-set: 1
Reviewer: Gerrit User 1044446 <1044446@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Presubmit-Ready=0