Add support for multiple children within a DrawNode.
Modified LayoutNode to consume a set of children
that will not be drawn directly by the LayoutNode
and instead will be rendered by another ComponentNode.
This is done to support use cases where a DrawNode would
wrap multiple child ComponentNodes. In this case
the DrawNode's children will be added to the most immediate
LayoutNode for measurement/layout however, will be drawn
by it instead of the parent LayoutNode.

Test: N/A, proof of concept
Bug: N/A
Change-Id: Ief56a8fa471fe32a9bd60b9ede324ccc227aa4be