Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5: Presubmit-Ready+1
> 
> Hi, Alan!
> How do I find out the reason for the failure of the test?
> When I click on the link, the login window appears. I do not have a login, so I can not see the reason why the test fails.
> Best regards,
> Sergey

sorry we don't have the setup to share them yet.
Here is the error:

/usr/local/google/buildbot/src/android/androidx-master-dev/frameworks/support/swiperefreshlayout/src/main/java/androidx/swiperefreshlayout/widget/SwipeRefreshLayout.java:266: Error: Access to private constructor of class SavedState requires synthetic accessor [SyntheticAccessor]
return new SavedState(in);
~~~~~~~~~~~~~~~~~~

Explanation for issues of type "SyntheticAccessor":
A private inner class which is accessed from the outer class will force the
compiler to insert a synthetic accessor; this means that you are causing
extra overhead. This is not important in small projects, but is important
for large apps running up against the 64K method handle limit, and
especially for libraries where you want to make sure your library is as
small as possible for the cases where your library is used in an app
running up against the 64K limit.

Patch-set: 5
Reviewer: Gerrit User 1060337 <1060337@85c56323-6fa9-3386-8a01-6480fb634889>