Update patch set 2
Patch Set 2: Code-Review+2

I'm OK with getting this in for now, as we should make sure Checkbox and other components work properly. However, this is a symptom to a deeper issue and this is not a proper fix for it. We should not mark the bug as fixed.

Patch-set: 2
Reviewer: Gerrit User 1493285 <1493285@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2