Update patch set 3
Patch Set 3:

(3 comments)

Thanks! I really like the move of the identity map out of FrameManager and make it more general. The code reads much better with the changes you made

This also fixes some "Not in a frame" exceptions as it framed objects throw when `hashCode` is called on them in a commit observer.

However, I would like to maintain the deduplication of calls to scope invalidate as it they can be expensive.

Patch-set: 3
Reviewer: Gerrit User 1048725 <1048725@85c56323-6fa9-3386-8a01-6480fb634889>