Ensure parent NavGraphs are added to Navigator
When using navigate via deeplink across graphs, we add the appropriate
parents to the navController back stack, but we do not add any parents
to the navigation navigator back stack. This causes an issue when we pop
a destination that was navigated to via deep link, where an extra
navGraph is removed from the navigation navigator back stack.
Eventually, the navigation navigator gets to the root and it is
impossible to return back from the deep link destination.

This change ensures that when parents are added to the
navController back stack, we also add them to the navigation navigator
back stack.

Test: New test, navigation tests
BUG: 132509387
Change-Id: I811282e359c0bc2ee04fd8aba4dfc6c9d50855f6