Update patch set 6
Patch Set 6:

> Patch Set 6:
> Test seems to have failed on the CI (it did pass locally, i just tried). Any ideas why?

Most likely, on the local machine tests are performed sequentially, while on the CI tests are performed in parallel.

The configuration change occurred while the code was being executed in the testStartAndStopRefreshing() method, in particular, during the execution of the code inside the loop: for (int i = 0; i < 5; i++) {...} at the second or third iteration, since the string 'Set SwipeRefreshLayout refreshing state' occurs three times in LogCat.

While the testStartAndStopRefreshing() method was run, the method testRefreshStatePersists() was run in parallel.

The configuration change resulted from the execution of a line of code: onView(isRoot()).Perform(orientationLandscape()), which is in the testRefreshStatePersists() method. This is reflected in LogCat, there is a line: "Device reconfigured: id = 8, name = 'synaptics_dsxv26', size 1080x1920, orientation 0, mode 1, display id 0".

Patch-set: 6
Reviewer: Gerrit User 1497627 <1497627@85c56323-6fa9-3386-8a01-6480fb634889>