Update patch set 3
Patch Set 3:

Discussed with Andrey and he pointed out that by drawing to a Bitmap we are losing hardware acceleration, but this might be alright for the moment as this code will probably be temporary until we switch to using RenderNodes. At that point it would be great (and achievable AFAWCT) to preserve the predictability of the measure/layout/draw passes enforced by this CL. What do you think?

Patch-set: 3
Reviewer: Gerrit User 1292888 <1292888@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0