Make image keyboard API backport compatible with all previous impls
This is a follow up CL to our recent attempt [1] to work around an
unintentional protocol breakage in image keyboard API on API<25 that
was accidentally introduced in AndroidX 1.0.0 [2].

Basically there are three versions of image keyboard API backports.
 - legacy support lib
 - AndroidX 1.0.0
 - upcoming Android X 1.1.0

As a result, there are 9 cases to consider depending on which version
is used in the app and which version is used by the IME.

With this CL, Android X 1.1.0 version becomes compatible to all the
versions above.  This should make our life much easier, because
regardless of whether it's an app or IME, apps/IMEs becomes compatible
with other IMEs/apps that were built with whatever version of
libraries above just by upgrading to AndroidX 1.1.0.

 [1]: I42a4a1a215a00d42fca62bec31e52b0947d1de74
      08a73da29759682576450ea7c31ad49722265f2c
 [2]: I11a047324832801555673dac45ec1d6590a6338b
      b31c3281d870e9abb673db239234d580dcc4feff

Bug: 129349688
Test: unittest
Change-Id: I6cc9498151a09c4f62435e4dd07def63c223e140